Add a new ProcessorMap class
This class is a simplified version of ProcessorSelector: instead of relying on
predicates to select the processor to choose, it relies on a Map, whose keys are
the result of the computation (via a Function) of the input.

As for ProcessorSelector, there is a(n optional) default processor to use if no
such key exists.

Now, it needs testing.