New abstract classes {AsJson},ValueHolder<T>
In quite a few cases, what is wanted as inputs and outputs of processors are
simple wrappers around values.

Create two dedicated abstract classes: one for a generic value, of whatever
type, and the other for values implementing AsJson.

This class, as for all processor inputs and outputs, implements MessageProvider.