JsonPatchOperationTest: remove test on notSame() :(
Unfortunately, some nodes are perfectly legal to be the same. This is the case
for NullNode, for instance.

Which means extra care is needed in the code itself. Or that a custom assert
needs to be written for testing whether the node is null, or a boolean, or...

Or maybe just create a mutable tree class of ours.