First foray into DI with Guice
Two primitive modules demonstrating DI used on SyntaxProcessor.

Dagger was also considered at one point, but unfortunately its @Module interface
is not @Inherited... Speaking with one of the authors on IRC, he said it was to
"encourage composition over inheritance". Well, yes, but here inheritance is
perfectly legitimate.

Too bad... Guice it is.